Moved login.dart to Pages folder
